python: add tests to verify every SVD can be successfully parsed
This code generates a test case (that can be picked up by nose) for
every MCU that is in tree.  The tests will pass if no exception
is raised and a value is returned.  Beyond that, there isn't too
much more we can verify against without having special knowledge.

This found a few bugs, so it certainly seems to have some merit.