core: cleanly support inheritance and derivation
This change moves to a dynamic model for resolving at runtime
derivation (derivedFrom) and inheritance (for specified keys).  This
is cleaner than the second-pass approach previously performed that
required additional funciton calls.

In addition, this commit adds a few missing attributes on fields,
registers, and peripherals.  Register arrays are now always populated
and the original register arrays are no longer present in the tree
(they are always expanded out).

Signed-off-by: Paul Osborne <osbpau@gmail.com>