core: fix to_dict()/json generation and add test
This test will probably fail on... most changes as it contains basically
everything.  That being said, I think it is useful for showing how
changes impact the generated data structure.  Just uncomment the
one line temporarily to regenerate and then examine the diff to
see if the change to the tree makes sense.

Signed-off-by: Paul Osborne <osbpau@gmail.com>