core: drop svd bundling behavior and prep 0.5
The former package was very large and I believe it makes most sense to
move away from this distribution model.  This commit drops the
previously supported pkgresources based approach (which is now deprecated
as well) in favor of requiring the caller point the parser at an SVD
explicitly.

A change to bump the version is included here.

Signed-off-by: Paul Osborne <osbpau@gmail.com>