format display_name with dim_indices
According to the SVD specification:
https://www.keil.com/pack/doc/CMSIS/SVD/html/elem_registers.html#elem_register,
the 'displayName' child element of the register element can also take
advantage of the '%s' syntax. The only difference with 'name' is that
'displayName' is optional.
This patch duplicates the logic used by 'name' for
'displayName'. Because it is not clear if 'displayName' must contain
'%s' let's handle any error.